Split server ishell and client ishell into seperate files. This is much easier to read and makes much more sense then having all commands in a single. This also prevents the conflit between the two `runCommand` functions and eliminates the need for runcmd.go
