--- torat_server/runcmd.go
@@ -1,25 +1,0 @@
-package server
-
-import (
-	"strings"
-
-	"github.com/abiosoft/ishell"
-	"github.com/lu4p/ToRat/models"
-)
-
-func (client *activeClient) runCommand(c *ishell.Context) {
-	command := strings.Join(c.Args, " ")
-	var r string
-	args := models.Cmd{
-		Cmd:        command,
-		Powershell: false,
-	}
-
-	err := client.RPC.Call("API.RunCmd", args, &r)
-	if err != nil {
-		c.Println(yellow("["+client.Client.Name+"] ") + red("[!] Bad or unkown command!"))
-		c.Println(yellow("["+client.Client.Name+"] ") + red("[!] ", err))
-	}
-
-	c.Println(r)
-}
